fix: ValueError in interp_solution_on_new_mesh
ValueError: could not broadcast input array from shape (160,) into shape (159,)

This bug is due to an inconsistency between the computed number of nodes on the
new mesh and the lenght of array zcoord_new. In particular, the number of nodes
is computed correclty while array zcoord_new is constructed in a wrong way
because of an error in the coarsening procedure.
This error is related to the fact that, in functin coarse_mesh, a node was
appended rather than replaced in the if statements
elif node_flag_new == False and node_flag_old and
elif node_flag_new == False and node_flag_old == False.
In practice the soft node representing the starting point of the element that
should have been removed to coarsen the mesh by means of a replacement with
the node from the old mesh, was not removed so the function fails to coarsen
the mesh.
This commit fix this bug and explains why the correction is made in function
coarse_mesh.

modified:   utility_functions/adaptive_mesh_functions.py