fix: adapt time step not working with adapt mesh
This silent bug is due to the fact that array SYSVAR_old is not correctly
updated in the case of adaptive mesh if the mesh between two consecutive
thermal hydraulic time step is not modified (i.e. no quench fronts are
detected).

To fix this error the the following methods and functions were modified:

    * class Conductor module conductor.py
        * initialization
        * update_dict_step_on_static_mesh
    * module adaptive_mesh_functions.py
        * adaptive_mesh

Method initialization

    Saved an hard copy of the thermal-hydraulic problem solution (SYSVAR) at
    the initialization in key SYSVAR_old of attribute dictionary dict_Step.
    This necessary to correctly evaluate the variation of the solution in
    time at the first thermal hydraulic time step. For all the other thermal
    hydraulic time step this update is managed by methods
    self.update_dict_step_on_static_mesh and self.interp_solution_on_new_mesh.

Method update_dict_step_on_static_mesh

    Updated the method in order to be used also if mesh is adaptive but it is
    not modified between two consecutive time steps. In particular, the check
    on the kind of mesh is removed from the method in order to be used regardless
    of the value of the flag ITYMSH.

Function adaptive_mesh

    If the mesh is not adapted (no quench fronts detected) method
    update_dict_step_on_static_mesh is called to update SYSVAR_old and SYSLOD
    without interpolations (as if the mesh is static).

It is defined as silend bug because it an issue that does not cause the
abrupt end of the simulation raising an error.

Class: Conductor
modified:   conductor.py
modified:   utility_functions/adaptive_mesh_functions.py