feat: new funciton update_time_evolution_io
New function that updates the dictionary time_evol_io that stores the time
evolutions of the property at the inlet and at the outlet. Since these
coordinates are fixed and correspont to indez 0 and -1, there is no need to
carry out interpolation to evaluate the values. This function can be used
only with instances of class FluidComponent.

modified:   utility_functions/output.py