feat: neew function save_te_on_file_io
New function that saves the time evolution of selectet variables at given the
inlet and at the outletof the conductor channel. If arrays have reached a
lenght equal to CHUNK_SIZE, their values are stored in appropriate files and the
conten of the array is cleared (new empty list are created) to perform a new
storage cycle. This allows to reduce the number of times the code reads and
writes files.
This function is introduced as a special case of function save_te_on_file
because in this case the keys of dictionaries t_evol are not the coordinates
but directly the properti names, as opposed to all the ther cases where
another nested dictionary is exploited.

modified:   utility_functions/output.py