chore: replace call to fun save_simulation_time
In methods conductor_initialization and conductor_solution, call to function
save_simulation_time is replaced with the new functions that deals with the
saving of time evolutions save_time_evolution_init and save_time_evolution.
The former initializes the data strucure and the output files used to save the
time evolutions, the latter acctually carries out the saving and storage in
output files.

Class: Simulation
modified:   simulation.py