feat: new function force_min_time_step
New function that forces the time step to the minimum value when the simulation
get closer to an event, to avoid strong discontinuity in the solution due to
switching on/off heating or changing the value of the current. This is forced
when, with the present time step lenght, the number of time steps necessary to
reach the time of the event is smaller or equal than 2. This value is chosen to
avoid forcing the minimum time step lenght too early, i.e. when the distance to
the event is smaller or equal than the maximum time step lenght.

modified:   utility_functions/transient_solution_functions.py