chore: use t_step_min before the event
Updated method conductor_solution of class Simulation to manipulate the
thermal hydraulic time step in order to use the minimum time step before the
beginning of an event. This is necessary to limit the issue of overshooting and
undershooting the thermal hydraulic solution (e.g. when there is a localized
square wave heating).

In order to achieve this goal, the next steps are carried out:
    * called function force_min_time_step in method conductor_solution;
    * reorganized the order of the call of functions force_time_step and
    time_and_event_synchronization. Specifically, the order in which these
    functions are called is inverted wrt previous commits.

This strategy is checked against two test cases and seems to work properly in
both of them.

Class: Simulation
modified:   simulation.py