chore: update initializaion of attr num_step_save
Updated the initializaion of attribute self.num_step_save of class Conductor in
method self.__initialize_attributes.
This array is now filled with negative integer values from -N to -1 where N is
the lenght of the array.

This is done to avoid overwriting the temporary file with the spatial
distributions characterized by (0) in its name in case of issues in the saving
of spatial distributions at user defined times. Moreover in this way the number
of missed saving could be identified from the number of temporari files
characterized by (-n) with n the index of the missed time for the saving from
sheet Spatial_distribution of input file conductor_diagnostic.xlsx after
sorting in ascending order. Counting starts from 1 and not 0!!

Class: Conductor
modified:   conductor.py