fix: ValueError in fun current_sharing_temperature_re123
ValueError: operands could not be broadcast together with shapes (151,) (217,)

This error is due to an incorrect check of flags IOP_MODE of the generic
SolidComponent instance and I0_OP_MODE of the conductor in method
self.__check_current_mode of class SolidComponent.
Specifically, the error occurs when I0_OP_MODE = None and IOP_MODE = 0 as in
this case the method does not update the arrays of the currrent (it only does
it at the initialization setting their vale to 0).
Since the array current_along is not updated, the evaluation the current
density lead to an array of inconsistend dimension that is passed to function
current_sharing_temperature_re123, from which the error is rised.

This fix consists in exploiting the new method __check_current_mode of class
Conductor to update the value of flag IOP_MODE of SolidComponent instances,
as it returns None if flag I0_OP_MODE = None.

Another possible fix to the bug is to correctly set flag IOP_MODE = None in the
input file conductor_operation.xlsx.

Class: Conductor
modified:   conductor.py