fix: ValueError in function save_conductor_sd
ValueError: All arrays must be of the same length

This bug is due to the not consistent shape of the arrays stored in each
dictionary in attribute self._store_sd_node of class Conductor in keyword
t_save.
Specifically, only the interfaces characterized by a variable htc had the
correct shape as they where interpolated in space on the new mesh (if adaptive).
Instead, the interfaces characterized by costant htc were not reshaped on the
new mesh size (if adaptive); therefore their length was always equal to the
initial number of nodes. This brings to the error and explains why the fix is
done in method self.store_interp_spatial_distributions of class Conductor.

Class: Conductor
modified:   conductor.py