fix: IndexError in function build_svec
IndexError: index 1 is out of bounds for axis 1 with size 1

This bug is related to the removal of key "JHTFLX" from attribute dict_node_pt
for all solid components.
The bug appears only with the test case 3 as it is the only case where there is
no localized heating on a solid component, so that Q1 and Q2 may not have the
expected shape (N_node, M) whit M = 2 if BE or CN are used and M = 4 if AM$ is
used to integrate in time.

To fix this bug the following changes are done in methods of class Conductor:

    1. in method initialization, the initialization of attribute
    jacket.radiative_heat_env with shape (N_nod, 1) is
    removed as it is replaced by the calling of method
    jacket._init_radiative_source_therm_env.
    2. In method __initialize_heat_source_nodal_pt_th called method
    jacket._init_radiative_source_therm_env to initialize with the correct
    shape array jacket.radiative_heat_env for all jacket component.
    3. In method __build_heat_source_nodal_pt updated the call of method
    jacket._radiative_source_therm_env to account for the return of the
    method (before the method updated the array inplace).

Need to implement a more efficient way to update array
jacket.radiative_heat_env for those components that do not have an
interface with the environment in the case of adaptive mesh. At the time
being an interpolation is performed.

Class: Conductor
modified:   conductor.py