fix: ValueError in method post_processing
ValueError: operands could not be broadcast together with shapes (549,) (544,)

This bug is due to the fact that the code computes a new time step and a new
grid (if adaptive and if requirements are satisfied) also when the end time of
the simulation is reached. This is not desired as in the last time step the
simulation should only compute the last solution and carry out the post
processing as prescribed.

The bug is fixed checking the simulation time and computing a new time step and
eventually adapting the grid only for simulation time that are lower than the
simulation end time.

Class: Simulation
modified:   simulation.py