fix: InterpolationError in store_interp_spatial_distributions
Removed interpolation of zcoord_old and zcoord_gauss_old on the index of the
elements in of the array with the spatial distribution as this led to a wrong
representation of the spatial discretization due to the incorrect removal of
the hard nodes of the mesh.
The hard nodes are removed only from arrays stored in attribute
self.store_sd_node of class Conductor, while in the actual array with the
spatial discretization the hard nodes are correctly handled by the algorithm
of the adapitive mesh.

This bug does not cause the crashing of the simulation but lead to an
incorrect representation of the spatial distribution.

The bug is identified with the set of input files associated to case test n4,
based on the VIPER like cable.

modified:   conductor.py
modified:   properties_of_materials/rare_earth_123.py