fix: ValueError in fun check_repeated_headings
ValueError: ERROR! Different objects of the same kind (Z_JACKET) can not have
the same identifier.

This bug appears if user defines more than one conductor or components with _0
in their identifier. For instance CHAN_1 CHAN_0 CHAN_0.
Althoug it is not allowed to have conductors or components with the same
identifier, allowing multiple identifier with _0 could be useful to not lose
informations in input files when removing some components that may not be
useful in that specific simulation. Indeed, conductor or components
characterized by _0 in theri identifier are not instantiated in the simulation
so they will not contribute to the outcome of the simulation.
It is therefore reasonable to exclude those components in the check of repeated
headings, increasing the flexibility of the input files.

Before this fix, the only option left to the user to achieve the same result
was to create a copy of the input file and delete the columns with the not
necessary conductors or components.

Bug found when trying to recicle input file of test case 3, removing jackets
from 3 to 6 replacing their number with 0.

modified:   utility_functions/auxiliary_functions.py