fix: ValueError in function assemble_syslod
ValueError: operands could not be broadcast together with shapes (6,) (12,) (6,)

This bug is due to the not update of array self.dict_Step["SYSLOD"] on the new
mesh at the firs time step due to the if statement "if self.cond_num_step > 1"
in method self.update_syslod_on_new_mesh of class Concutor.

This check is not correct since the adaptive mesh is called after the call to
the step function, so, from the theorical point of view the time step is
already the next one. Therefore it is important to remove that if statement in
order to correctly interpolate array self.dict_Step["SYSLOD"] also in cases the
mesh is updated after the first time step.

This bug occurred with a set of input file used to test CASE 3bis with
parameters such that mesh adaptation is performed right after the frist time
step after the initialization.

With the same rationale also method self.update_dict_step_on_static_mesh of
class Conductor is corrected.

Fix in method
    * self.update_syslod_on_new_mesh
    * self.update_dict_step_on_static_mesh

Class: Conductor
modified:   conductor.py