fix: no time step relaxation after min time step
The adaptive time step did not increases afeter it reaches the minium value
due to the incorrect update of keyword "SYSVAR_old" of attribute
dictionary self.dict_Step of class Conductor.
The value in this keyword is used to compute the eighenvalues used to evaluate
the adaptive time step.
The value is not correctly updated because inserted into a branch of an if that
raises a value error in method self.update_dict_step_on_static_mesh of class
Conductor.

This is why method this fix involves method self.update_dict_step_on_static_mesh.

The bug is found for any set of input files that uses IADAPTIME = 1 or 2 and BE
or CN as method for the integration in time.

Class: Conductor
modified:   conductor.py