fix: metadata files saved in wrong folder
Due to the removal of the GUI, the metadata version of the input files is
saved in the wrong folder, that is the same where the input files are loaded.
This causes the not correct working of the code after the first run.

In the yaml configuration file is added a keyword "input_folder" that requires
user to provide the name of the folder where the input files are stored.

In module opensc2.py to the constructor of the class Simulation is passed the
whole dictionary io_paths and not only keyword "input".

In class Simulation a new attribute input_file_folder is defined to get the
value assigned to "input_folder" and used to correctly create the folder where
to save the input files as metadata in the output path.

Modified following methods of class Simulation:
    * __init__
    * simulation_folder_manager

Class: Simulation
modified:   opensc2.py
modified:   simulation.py