refactor: retrun values in function fixed_value
The function not longer modifies inplace the attributes of class
Conductor, but returnd their value in a tuple collecting the aliases
introduced in the function. The attributes are:

    * fixed_potential_value
    * fixed_potential_index
    * electric_known_term_vector
    * electric_stiffness_matrix

Docstring is updated accordinlgy.

This refactorin to functon fixed_value is useful to avoid the change of shape
of attribute electric_stiffness_matrix; in this way it is not longer necessary
to re-allocate the correct shape at each electric time step, meaning that
memory can be allocated (in the worst case scenario) at each new
thermal-hydraulic time step. In this way it is possible to remove the memory
allocation from method __build_electric_stiffness_matrix and move to a
dedicated method, so that it is possible to have a method that calls the method
__build only for those matrices thas must be uptated without wasting time in
memory allocation.
This commit is only the first step to this goal of optimization of the electric
method, function electric_transient_solution with the aim to remove not necessary
calls to method electric_preprocessing of class Conductor in view of the introduction
of the adaptive mesh.

modified:   utility_functions/electric_auxiliary_functions.py