refactor: electric solution functions
Update call to functon fixed_value in functions electric_steady_state_solution
and electric_transient_solution to account for the new return of that function.

Note that attribute electric_stiffness_matrix of class Conductor is not updated
with the returned stiff_mat, that is assigned to a local variable
(stiffnes_matrix) and passed to the scipy solver.
In this way the shape of attribute electric_stiffness_matrix is not changed and
there is no need to memory re-allocation at each electric time step.

This commit is the second step to the optimization of the electric method,
function electric_transient_solution with the aim to remove not necessary
calls to method electric_preprocessing of class Conductor, in view of the
introduction of the adaptive mesh.

modified:   utility_functions/electric_auxiliary_functions.py