refactor: funcition electric_steady_state_solution
Replaced call to method conductor.electric_preprocessing with the more
specific and coherent method conductor.electric_preprocessing_initialization.
This is done because at the time being the flow chart of the electic model
assumes that at the beginning of the problem (conductor.cond_num_step = 0), the
seady state version of the electric problem is solved. Then, for all the other
steps, function electric_transient_solution is used.
Thus, this is the correct function where to call the method that initializes
ndarrays and data structure used in the electric problem.

modified:   utility_functions/electric_auxiliary_functions.py