fix: ValueError in method post_processing
ValueError: operands could not be broadcast together with shapes (137,) (136,)

Bug due a floating point representation affecting the if statement
if simulation.simulation_time[-1] < simulation.transient_input["TEND"]

If TEND = 5 with costant dt = 0.1, the last time step (that should be 5 s) is
represented as 4.99999999999993 so that the if above if statement is still true
even though it shoudl be false. So the code enters the first branch of the if
statement and not the second one as expected performing wrong operations that
lead to the bug.

Fixed exploiting function np.iscolose.

Class: Simulation
modified:   simulation.py